move generator macros to provider
